+++ src/main/java/com/leets/X/domain/user/service/LoginStatus.java
@@ -1,4 +1,4 @@
-package com.leets.X.domain.user.domain.enums;
+package com.leets.X.domain.user.service;
 
 public enum LoginStatus {
     LOGIN, REGISTER
